FKS-1028 reconfig testfile
